centrala/requirements.txt
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
Flask==2.1.1
Werkzeug==2.1.0
Pillow==9.0.1
requests==2.27.1
pandas==1.4.1
numpy==1.22.3
tqdm==4.63.1
netifaces==0.11.0